Split tests/lifecycle_regression.rs into lifecycle_{claim,recovery,terminal}.rs
